Rename virtool_workflow.analysis.subtraction
Change from `virtool_workflow.analysis.subtraction` to `virtool_workflow.analysis.subtractions`.
