+++ docs/reference.rst
@@ -28,14 +28,14 @@ API Reference
 
 
 ``virtool_workflow.analysis.library_types``
-==================================
+===========================================
 
 .. automodule:: virtool_workflow.analysis.library_types
     :members:
 
 
 ``virtool_workflow.analysis.reads``
-==================================
+===================================
 
 .. automodule:: virtool_workflow.analysis.reads
     :members:
@@ -46,7 +46,7 @@ API Reference
 .. automodule:: virtool_workflow.analysis.sample
     :members:
 
-``virtool_workflow.analysis.subtraction``
+``virtool_workflow.analysis.subtractions``
 ==========================================
 
 .. automodule:: virtool_workflow.analysis.subtractions
