+++ virtool_workflow/analysis/fixtures.py
@@ -1,5 +1,5 @@
 from virtool_workflow.analysis.sample import sample, paired
-from virtool_workflow.analysis.subtraction import subtractions
+from virtool_workflow.analysis.subtractions import subtractions
 from virtool_workflow.analysis.hmms import hmms
 from virtool_workflow.analysis.analysis import analysis
 from virtool_workflow.analysis.indexes import indexes
